requirements.txt

flask
flask_sqlalchemy
matplotlib
pillow
numpy
